Update cart, add BadgeIcon above CartIcon to show how many sneakers were added to the cart
* Make cart items stay in the cart when reloading the page
* Remove items prop from Cart component
* Add badge above cart
* Make the total sum of the cart change without reloading the page when adding a new pair of sneakers

Place `CartIcon` and `BadgeIcon` classes in a div with `CartBadgeGroup` class to make the badge increase in size when hovering over the cart icon.
In `Cart/index.tsx` add `useEffect` function and move `forEach` loop into it. Add `dispatch` function in `useEffect` to set items and sum in the cart when
the page is loaded. Add `dispatch` to the list of dependencies of the `useEffect` function.
Rename `IProps` from types/icons to IconProps. Move `mainItemsSlug` in Root/index.tsx to settings.
Rename `Items` interface to `State`, add `sum` as a value to it. Add two new functions to `ActionTypes` enum: `setItems` and `setSum`.