Create `store/stateFunctions` for managing the app's state and move all endpoints to `settings/endpoints`
Add `onClickRemoveButton` in `Cart/index.tsx`. Add `onClick` function to `RemoveOrCloseIcon`.
Add `HttpStatus` enum for managing responses' status codes. It is necessary to know if the
server returns an error, so I made the app check the status code of a response and throw an
error if it returns something starting with a 4 or 5.